% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hClust_llr.R
\name{hClust_llr}
\alias{hClust_llr}
\title{Hierarchical Clustering based on log-likelihood ratio}
\usage{
hClust_llr(X, ZI = TRUE, reorder = TRUE)
}
\arguments{
\item{X}{matrix n x p with observations (i.e. individuals) in rows and variables in columns}

\item{ZI}{boolean, indicating whether using zero-inflated Gaussian model (if TRUE) or
just gaussian model (if FALSE), default is TRUE}

\item{reorder}{boolean, indicating whether individuals should be reorder or not
(for heatmap aesthetic), if TRUE, use a re-implementation of \code{\link[gclus]{reorder.hclust}}}
}
\value{
an object of type hclust, see \code{\link[stats]{hclust}}
\item{diss}{the n x n dissimilarity matrix between individuals}
\item{}{(not included in \code{hclust} object)}
\item{height}{\code{\link[stats]{hclust}}}
\item{merge}{\code{\link[stats]{hclust}}}
\item{order}{\code{\link[stats]{hclust}}}
\item{labels}{vector of n individual labels}
}
\description{
\code{hClust_llr} returns an object of type \code{hclust},
results of the clustering on individuals, using the dissimilarity
based on log-likelihood ratio test.
}
\details{
Algorithm to process hierarchical clustering on individuals
(from a matrix individuals x variables). The dissimilarity used
in the observation space is the log-likelihood ratio, i.e.
D(i,j) = log f(Y_i) + log f(Y_j) - log f(Y_i U Y_j)
where Y_i is the vector of observations from individual i.
The statistical model defining the considered likelihood is
standard Gaussian or Zero-Inflated Gaussian (i.e. a Gaussian-Bernoulli
mixture). The agglomeration method is also the log-likelihood ratio,
however between clusters in this case.

Wrapper for Cpp function
}
\author{
Ghislain Durif, \email{ghislain.durif@univ-lyon1.fr}
Franck Picard, \email{franck.picard@univ-lyon1.fr}
}
\seealso{
\code{\link{plot_heatmap}}
}

