% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Heatmap of Hierarchical Clustering based on log-likelihood ratio}
\usage{
plot_heatmap(diss_matrix, clustering, rSideCol = NULL, cSideCol = NULL,
  title = NULL)
}
\arguments{
\item{diss_matrix}{dissimilarity matrix between individuals (n x n) used for clustering}

\item{clustering}{an object of type hclust, generated by \code{\link{hClust_llr}}}

\item{rSideCol}{a vector of color to add on the row side of the heatmap (default is NULL)}

\item{cSideCol}{a vector of color to add on the column side of the heatmap (default is NULL)}

\item{title}{a title (default is NULL)}
}
\description{
\code{plot_heatmap} draws the heatmap after the clustering from \code{\link{hClust_llr}}
}
\details{
Plot the result from \code{\link{hClust_llr}}
}
\author{
Ghislain Durif, \email{ghislain.durif@univ-lyon1.fr}
}
\seealso{
\code{\link{hClust_llr}}
}

