\name{ziLRhClust-package}
\alias{ziLRhClust-package}
\alias{ziLRhClust}
\docType{package}
\title{
\packageTitle{ziLRhClust}
}
\description{
\packageDescription{ziLRhClust}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ziLRhClust}
\packageIndices{ziLRhClust}
%%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ziLRhClust}

Maintainer: \packageMaintainer{ziLRhClust}
}
\references{
~~ Literature or other references for background information ~~
}
%%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
%%\keyword{ package }
\seealso{\code{\link[stats]{hclust}}}
%%~~ Optional links to other man pages, e.g. ~~
%%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~

%%\examples{
%% ~~ simple examples of the most important functions ~~
%%}
